Changed the location of some variables
Did it so that the program can update fields it needs when a photo is
done. I also made the encrypt and decrypt methods threads so the main
window would be responsive. I want to add a checkbox so the user can say
yes or no to live updating (it would improve performance)